MF-031 Refactor utilService and extract StationStyles
